Enhance CI containers management
- permit scripts to be executed from any path;
- handle containers init in a unique script;
- exec docker commands using docker-compose exec;
- do healthcheck on all containers listed in compose files;
- try to clean old DB and fixtures before recreating them;
- clean only containers listed in compose files.

+++ .github/actions/init_install-dependencies.sh
@@ -4,12 +4,18 @@ PHP_MAJOR_VERSION="$(echo $PHP_VERSION | cut -d '.' -f 1,2)"
 
 # Validate composer config
 composer validate --strict
-if ! [[ "$PHP_MAJOR_VERSION" == "8.0" ]]; then composer check-platform-reqs; fi
+if ! [[ "$PHP_MAJOR_VERSION" == "8.0" ]]; then
+  composer check-platform-reqs;
+fi
 
-# Remove composer.lock and unset platform to test with dependencies matching PHP version
-rm composer.lock
-composer config --unset platform
+if [[ "$IGNORE_LOCK_FILES" = true ]]; then
+  # Remove composer.lock and unset platform to test with dependencies matching PHP version
+  rm composer.lock
+  composer config --unset platform
+fi
 
 # Install dependencies
-if [[ "$PHP_MAJOR_VERSION" == "8.0" ]]; then export COMPOSER_ADD_OPTS=--ignore-platform-reqs; fi
+if [[ "$PHP_MAJOR_VERSION" == "8.0" ]]; then
+  COMPOSER_ADD_OPTS=--ignore-platform-reqs;
+fi
 bin/console dependencies install --composer-options="$COMPOSER_ADD_OPTS --prefer-dist --no-progress"
